+++ src/kestrel/session.py
@@ -52,7 +52,6 @@ import shutil
 import uuid
 import logging
 import re
-import toml
 import time
 import math
 import lark
@@ -60,7 +59,6 @@ from datetime import datetime
 
 from kestrel.exceptions import (
     KestrelSyntaxError,
-    NoValidConfiguration,
     InvalidStixPattern,
 )
 from kestrel.syntax.parser import get_all_input_var_names
@@ -77,7 +75,8 @@ from kestrel.codegen.display import DisplayBlockSummary
 from kestrel.codegen.summary import gen_variable_summary
 from firepit import get_storage
 from firepit.exceptions import StixPatternError
-from kestrel.utils import set_current_working_directory, config_paths
+from kestrel.utils import set_current_working_directory
+from kestrel.config import load_config
 from kestrel.datasource import DataSourceManager
 from kestrel.analytics import AnalyticsManager
 
@@ -172,7 +171,7 @@ class Session(object):
             f"Establish session with session_id: {session_id}, runtime_dir: {runtime_dir}, store_path:{store_path}, debug_mode:{debug_mode}"
         )
 
-        self.config = self._load_configuration()
+        self.config = load_config()
 
         if session_id:
             self.session_id = session_id
@@ -239,9 +238,7 @@ class Session(object):
         # {"var": VarStruct}
         self.symtable = {}
 
-        self.data_source_manager = DataSourceManager(
-            self.config["language"]["default_datasource_schema"]
-        )
+        self.data_source_manager = DataSourceManager(self.config)
         self.analytics_manager = AnalyticsManager(
             self.config["language"]["default_analytics_schema"]
         )
@@ -531,34 +528,6 @@ class Session(object):
     def _update_symbol_table(self, output_var_name, output_var_struct):
         self.symtable[output_var_name] = output_var_struct
         self.symtable[self.config["language"]["default_variable"]] = output_var_struct
-
-    def _load_configuration(self):
-
-        configs = []
-
-        for path in config_paths():
-            try:
-                configs.append(toml.load(path))
-                _logger.debug(f"Configuration file {path} loaded successfully.")
-            except FileNotFoundError:
-                _logger.debug(f"Configuration file {path} does not exist.")
-            except toml.decoder.TomlDecodeError:
-                _logger.debug(f"Invalid configuration file {path}.")
-
-        if not configs:
-            raise NoValidConfiguration
-        else:
-            config = configs.pop(0)
-            for c in configs:
-                for domain, mappings in c.items():
-                    if domain in config:
-                        config[domain].update(mappings)
-                    else:
-                        config[domain] = mappings
-
-        _logger.debug(f"Configuration loaded: {config}")
-
-        return config
 
     def _leave_exit_marker(self):
         exit_marker = os.path.join(