+++ src/kestrel_datasource_stixshifter/interface.py
@@ -1,30 +1,92 @@
-"""STIX Shifter data source package provides access to data sources via
-stix-shifter.
+"""The STIX shifter data source package provides access to data sources via
+`stix-shifter`_.
 
-Before use, need to install the target stix-shifter connector packages such as
-``stix-shifter-modules-carbonblack``.
+Before use, one need to install any target stix-shifter connector packages such
+as ``stix-shifter-modules-carbonblack``. Check all `avaliable connectors`_ and
+their `pypi packages <https://pypi.org/search/?q=stix-shifter-modules&o=>`_.
 
 The STIX Shifter interface can reach multiple data sources. The user needs to
-setup one *profile* per data source. The profile name will be used in the
-``FROM`` clause of the Kestrel ``GET`` command, e.g., ``newvar = GET entity-type
-FROM stixshifter://profilename WHERE ...``. Kestrel runtime will load the profile
-for the used profile from environment variables:
+provide one *profile* per data source. The profile name (case insensitive) will
+be used in the ``FROM`` clause of the Kestrel ``GET`` command, e.g., ``newvar =
+GET entity-type FROM stixshifter://profilename WHERE ...``. Kestrel runtime
+will load profiles from 3 places (the later will override the former):
 
-- ``STIXSHIFTER_PROFILENAME_CONNECTOR``: the STIX Shifter connector name, e.g., ``elastic_ecs``.
-- ``STIXSHIFTER_PROFILENAME_CONNECTION``: the STIX Shifter `connection <https://github.com/opencybersecurityalliance/stix-shifter/blob/master/OVERVIEW.md#connection>`_ object in JSON string.
-- ``STIXSHIFTER_PROFILENAME_CONFIG``: the STIX Shifter `configuration <https://github.com/opencybersecurityalliance/stix-shifter/blob/master/OVERVIEW.md#configuration>`_ object in JSON string.
+#. stix-shifter interface config file (only when a Kestrel session starts):
 
-Properties of profile name:
+    Put your profiles in the stix-shifter interface config file (YAML):
 
-- Not case sensitive, e.g., ``profileX`` in the Kestrel command will match
-  ``STIXSHIFTER_PROFILEX_...`` in environment variables.
-- Cannot contain ``_``.
+    - Default path: ``~/.config/kestrel/stixshifter.yaml``.
+    - A customized path specified in the environment variable ``KESTREL_STIXSHIFTER_CONFIG``.
+
+    Example of stix-shifter interface config file containing profiles:
+
+    .. code-block:: yaml
+
+        profiles:
+            host101:
+                connector: elastic_ecs
+                connection:
+                    host: elastic.securitylog.company.com
+                    port: 9200
+                    indices: host101
+                config:
+                    auth:
+                        id: VuaCfGcBCdbkQm-e5aOx
+                        api_key: ui2lp2axTNmsyakw9tvNnw
+            host102:
+                connector: qradar
+                connection:
+                    host: qradar.securitylog.company.com
+                    port: 443
+                config:
+                    auth:
+                        SEC: 123e4567-e89b-12d3-a456-426614174000
+            host103:
+                connector: cbcloud
+                connection:
+                    host: cbcloud.securitylog.company.com
+                    port: 443
+                config:
+                    auth:
+                        org-key: D5DQRHQP
+                        token: HT8EMI32DSIMAQ7DJM
+
+#. environment variables (only when a Kestrel session starts):
+
+    Three environment variables are required for each profile:
+
+    - ``STIXSHIFTER_PROFILENAME_CONNECTOR``: the STIX Shifter connector name,
+      e.g., ``elastic_ecs``.
+    - ``STIXSHIFTER_PROFILENAME_CONNECTION``: the STIX Shifter `connection
+      <https://github.com/opencybersecurityalliance/stix-shifter/blob/master/OVERVIEW.md#connection>`_
+      object in JSON string.
+    - ``STIXSHIFTER_PROFILENAME_CONFIG``: the STIX Shifter `configuration
+      <https://github.com/opencybersecurityalliance/stix-shifter/blob/master/OVERVIEW.md#configuration>`_
+      object in JSON string.
+
+    Example of environment variables for a profile:
+
+    .. code-block:: console
+
+        $ export STIXSHIFTER_HOST101_CONNECTOR=elastic_ecs
+        $ export STIXSHIFTER_HOST101_CONNECTION='{"host":"elastic.securitylog.company.com", "port":9200, "indices":"host101"}'
+        $ export STIXSHIFTER_HOST101_CONFIG='{"auth":{"id":"VuaCfGcBCdbkQm-e5aOx", "api_key":"ui2lp2axTNmsyakw9tvNnw"}}'
+
+#. any in-session edit through the ``CONFIG`` command.
+
+If you launch Kestrel in debug mode, stix-shifter debug mode is still not
+enabled by default. To record debug level logs of stix-shifter, create
+environment variable ``KESTREL_STIXSHIFTER_DEBUG`` with any value.
+
+.. _stix-shifter: https://github.com/opencybersecurityalliance/stix-shifter
+.. _avaliable connectors: https://github.com/opencybersecurityalliance/stix-shifter/blob/develop/OVERVIEW.md#available-connectors
 
 """
 
-import os
 import json
 import time
+import copy
+import logging
 
 from stix_shifter.stix_translation import stix_translation
 from stix_shifter.stix_transmission import stix_transmission
@@ -32,13 +94,15 @@ from stix_shifter.stix_transmission import stix_transmission
 from kestrel.utils import mkdtemp
 from kestrel.datasource import AbstractDataSourceInterface
 from kestrel.datasource import ReturnFromFile
-from kestrel.exceptions import (
-    InvalidDataSource,
-    DataSourceError,
-    DataSourceManagerInternalError,
+from kestrel.exceptions import DataSourceError, DataSourceManagerInternalError
+from kestrel_datasource_stixshifter.config import (
+    RETRIEVAL_BATCH_SIZE,
+    get_datasource_from_profiles,
+    load_profiles,
+    set_stixshifter_logging_level,
 )
 
-from kestrel_datasource_stixshifter.config import ENV_VAR_PREFIX, RETRIEVAL_BATCH_SIZE
+_logger = logging.getLogger(__name__)
 
 
 class StixShifterInterface(AbstractDataSourceInterface):
@@ -48,41 +112,41 @@ class StixShifterInterface(AbstractDataSourceInterface):
         return ["stixshifter"]
 
     @staticmethod
-    def list_data_sources():
+    def list_data_sources(config):
         """Get configured data sources from environment variable profiles."""
-        data_sources = []
-
-        env_vars = os.environ.keys()
-        stixshifter_vars = filter(lambda x: x.startswith(ENV_VAR_PREFIX), env_vars)
-        for evar in stixshifter_vars:
-            profile = evar.split("_")[1].lower()
-            if profile not in data_sources:
-                data_sources.append(profile)
-
+        if not config:
+            config["profiles"] = load_profiles()
+        data_sources = list(config["profiles"].keys())
         data_sources.sort()
-
         return data_sources
 
     @staticmethod
-    def query(uri, pattern, session_id=None):
+    def query(uri, pattern, session_id, config):
         """Query a stixshifter data source."""
         scheme, _, profile = uri.rpartition("://")
         profiles = profile.split(",")
+
+        if not config:
+            config["profiles"] = load_profiles()
 
         if scheme != "stixshifter":
             raise DataSourceManagerInternalError(
                 f"interface {__package__} should not process scheme {scheme}"
             )
 
+        set_stixshifter_logging_level()
+
         ingestdir = mkdtemp()
         query_id = ingestdir.name
         bundles = []
+        _logger.debug(f"prepare query with ID: {query_id}")
         for i, profile in enumerate(profiles):
-            (
-                connector_name,
-                connection_dict,
-                configuration_dict,
-            ) = StixShifterInterface._get_stixshifter_config(profile)
+            # STIX-shifter will alter the config objects, thus making them not reusable.
+            # So only give stix-shifter a copy of the configs.
+            # Check `modernize` functions in the `stix_shifter_utils` for details.
+            (connector_name, connection_dict, configuration_dict,) = map(
+                copy.deepcopy, get_datasource_from_profiles(profile, config["profiles"])
+            )
 
             data_path_striped = "".join(filter(str.isalnum, profile))
             ingestfile = ingestdir / f"{i}_{data_path_striped}.json"
@@ -104,6 +168,9 @@ class StixShifterInterface(AbstractDataSourceInterface):
                 raise DataSourceError(
                     f"STIX-shifter translation failed with message: {dsl['error']}"
                 )
+
+            _logger.debug(f"STIX pattern to query: {pattern}")
+            _logger.debug(f"translate results: {dsl}")
 
             # query results should be put together; when translated to STIX, the relation between them will remain
             connector_results = []
@@ -165,6 +232,8 @@ class StixShifterInterface(AbstractDataSourceInterface):
                         f"STIX-shifter transmission.query() failed with message: {stix_shifter_error_msg}"
                     )
 
+            _logger.debug("transmission succeeded, start translate back to STIX")
+
             stixbundle = translation.translate(
                 connector_name,
                 "results",
@@ -173,81 +242,9 @@ class StixShifterInterface(AbstractDataSourceInterface):
                 {},
             )
 
+            _logger.debug(f"dumping STIX bundles into file: {ingestfile}")
             with ingestfile.open("w") as ingest:
                 json.dump(stixbundle, ingest, indent=4)
             bundles.append(str(ingestfile.resolve()))
 
         return ReturnFromFile(query_id, bundles)
-
-    @staticmethod
-    def _get_stixshifter_config(profile_name):
-        profile_name = profile_name.upper()
-
-        env_conr_name = f"{ENV_VAR_PREFIX}{profile_name}_CONNECTOR"
-        connector_name = os.getenv(env_conr_name)
-        if not connector_name:
-            raise InvalidDataSource(
-                profile_name,
-                "stixshifter",
-                f"no {env_conr_name} environment variable found",
-            )
-        connector_name = connector_name.lower()
-
-        env_conn_name = f"{ENV_VAR_PREFIX}{profile_name}_CONNECTION"
-        connection = os.getenv(env_conn_name)
-        if not connection:
-            raise InvalidDataSource(
-                profile_name,
-                "stixshifter",
-                f"no {env_conn_name} environment variable found",
-            )
-
-        env_conf_name = f"{ENV_VAR_PREFIX}{profile_name}_CONFIG"
-        configuration = os.getenv(env_conf_name)
-        if not configuration:
-            raise InvalidDataSource(
-                profile_name,
-                "stixshifter",
-                f"no {env_conf_name} environment variable found",
-            )
-
-        try:
-            connection = json.loads(connection)
-        except json.decoder.JSONDecodeError:
-            raise InvalidDataSource(
-                profile_name,
-                "stixshifter",
-                f"invalid JSON in {env_conn_name} environment variable",
-            )
-
-        if "host" not in connection:
-            raise InvalidDataSource(
-                profile_name,
-                "stixshifter",
-                f'invalid {env_conn_name} environment variable: no "host" field',
-            )
-
-        if "port" not in connection and connector_name != "stix_bundle":
-            raise InvalidDataSource(
-                profile_name,
-                "stixshifter",
-                f'invalid {env_conn_name} environment variable: no "port" field',
-            )
-
-        try:
-            configuration = json.loads(configuration)
-        except json.decoder.JSONDecodeError:
-            raise InvalidDataSource(
-                profile_name,
-                "stixshifter",
-                f"invalid JSON in {env_conf_name} environment variable",
-            )
-
-        if "auth" not in configuration:
-            raise InvalidDataSource(
-                profile_name,
-                "stixshifter",
-                f'invalid {env_conf_name} environment variable: no "auth" field',
-            )
-
-        return connector_name, connection, configuration